+++ Project/DuckEngine/Game/src/GameScene.cpp
@@ -748,7 +748,7 @@ void GameScene::Update()
 		if (FadeOutSprite) {
 			FadeOutSprite->isVisible = true;
 			float fadeValue = (countdownTime / 3.0f) * 255.0f;
-			FadeOutSprite->color.a = (fadeValue >= 0.0f) ? static_cast<int>(fadeValue) : 0;  // Prevent negative values
+			FadeOutSprite->color.a = (fadeValue >= 0.0f) ? fadeValue : 0;  // Prevent negative values
 		}
 
 		return;  // Skip game logic until countdown is done
@@ -810,7 +810,7 @@ void GameScene::Update()
 				else {
 					float newVolume = TimeLeftSound->volume * (1.0f - fadeProgress);
 					SoundSystem::SetSoundVolume(TimeLeftSound->soundID[1], newVolume);
-					FadeOutSprite->color.a = static_cast<int>(fadeProgress * 255.0f);
+					FadeOutSprite->color.a = fadeProgress * 255.0f;
 				}
 			}
 		}
